+++ src/reducers/user.js
@@ -3,9 +3,6 @@ import * as UserActionTypes from '../constants/UserActionTypes';
 
 const initialState = {
     loginState: UserActionTypes.LOGGEDIN_CHECKING,
-    uid: '',
-    user: {},
-    displayName: 'guest'
 };
 
 
@@ -15,24 +12,21 @@ export default function user(state = initialState, action) {
     switch (action.type) {
 
         case UserActionTypes.LOGGEDIN_CHECKING:
-        console.log('UserActionTypes.LOGGEDIN_CHECKING');
-
+            // console.log('user.reducer: ', 'UserActionTypes.LOGGEDIN_CHECKING');
             return Object.assign({}, state, {
                 loginState: UserActionTypes.LOGGEDIN_CHECKING,
 
             });
 
         case UserActionTypes.LOGGEDIN:
-            console.log('UserActionTypes.LOGGEDIN');
+            // console.log('user.reducer: ', 'UserActionTypes.LOGGEDIN');
             return Object.assign({}, state, {
                 loginState: UserActionTypes.LOGGEDIN,
-                uid: action.uid,
-                user: action.user,
-                displayName: action.displayName
+                uid: action.uid
             });
 
         case UserActionTypes.LOGGEDOUT:
-            console.log('UserActionTypes.LOGGEDOUT');
+            // console.log('user.reducer: ', 'UserActionTypes.LOGGEDOUT');
             return Object.assign({}, state, {
                 ...initialState,
                 loginState: UserActionTypes.LOGGEDOUT,
@@ -40,17 +34,19 @@ export default function user(state = initialState, action) {
 
 
         case UserActionTypes.AUTHERROR:
+        // console.log('user.reducer: ', 'UserActionTypes.LOGGEDOUT');
+
             return Object.assign({}, state, {
                 loginState: UserActionTypes.AUTHERROR,
                 error: action.message
             });
 
 
-
         case UserActionTypes.UPDATE_USER_PROFILE:
-            console.log('UPDATE_USER_PROFILE');
+            // console.log('user.reducer: ','UserActionTypes.UPDATE_USER_PROFILE');
+            // console.log(action);
             return Object.assign({}, state, {
-                error: action.user
+                profile: action.profile
             });