+++ src/screens/navigationOnboarding.js
@@ -1,0 +1,23 @@
+import {StackNavigator} from 'react-navigation'
+
+// import Register from './screenRegister'
+import Login from './loginScreen'
+// import PwdForget from './screenPwdforget'
+// import Tour from './screenTour'
+
+const OnboardingNavigator = StackNavigator({
+    Login: {
+        screen: Login
+    },
+    // Register: {
+    //     screen: Register
+    // },
+    // PwdForgot: {
+    //     screen: PwdForgot
+    // },
+    // Tour: {
+    //     screen: Tour
+    // }
+}, {initialRouteName: 'Login'})
+
+export default OnboardingNavigator